refactor: move selected books buttons in separate component
