+++ src/components/booksList/BooksList.css
@@ -21,13 +21,6 @@
   width: 40%;
 }
 
-.selected-books-block {
-  display: flex;
-  justify-content: center;
-  align-items: baseline;
-  gap: 20px;
-}
-
 .book-checkbox {
   margin-right: 20px;
   width: 20px;
