+++ src/components/selectedBooksButtons/SelectedBooksButtons.css
@@ -1,0 +1,6 @@
+.selected-books-block {
+  display: flex;
+  justify-content: center;
+  align-items: baseline;
+  gap: 20px;
+}
