+++ src/store/selected-books/helpers.ts
@@ -1,0 +1,6 @@
+export function getSelectedMessage(selectedBooks: number) {
+  if (selectedBooks === 1) {
+    return `${selectedBooks} book is selected`;
+  }
+  return `${selectedBooks} books are selected`;
+}
